Remove PersonPrincipal. Now Person implements UserDetails interface.
