--- src/main/java/com/unvise/bankingsystemapp/domain/person/person/PersonPrincipal.java
@@ -1,51 +1,0 @@
-package com.unvise.bankingsystemapp.domain.person.person;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.Collection;
-
-@Getter
-@AllArgsConstructor
-public class PersonPrincipal implements UserDetails {
-
-    private Person person;
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return person.getRoles();
-    }
-
-    @Override
-    public String getPassword() {
-        return person.getPassword();
-    }
-
-    @Override
-    public String getUsername() {
-        return person.getEmail();
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return person.getStatus();
-    }
-
-}
